# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Borrowly API",
        "description": "API for Borrowly - Item Borrowing Platform",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/admin/dashboard": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get admin dashboard data",
                "operationId": "2c29ca3389addbd18db1fff0f925cfd3",
                "responses": {
                    "200": {
                        "description": "Dashboard data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List users",
                "operationId": "6df5675e4162a2cb784fd7bd54a30be7",
                "responses": {
                    "200": {
                        "description": "Users"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/users/{user}/role": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Change user role",
                "operationId": "a8cd75ca843c9dd44d8940099ee76ed1",
                "responses": {
                    "200": {
                        "description": "Role updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/users/{user}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete an unused user account",
                "operationId": "c53daedf3d979a9d6840f0aa25f1263a",
                "responses": {
                    "200": {
                        "description": "User deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/items": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List items for moderation",
                "operationId": "21f874624e2fee14ebd2b6ed7c64e4bc",
                "responses": {
                    "200": {
                        "description": "Items"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/items/{item}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete an item with no active borrowing",
                "operationId": "5ca8de4bb8fc303b268db04a3775f9f5",
                "responses": {
                    "200": {
                        "description": "Item deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/reports": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List reports",
                "operationId": "4c8be7d2e2de6d7f7663ae5a01152b70",
                "responses": {
                    "200": {
                        "description": "Reports"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/reports/{report}": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Resolve or dismiss a report",
                "operationId": "d8cebd35f07d803c307ca14e8f38098b",
                "responses": {
                    "200": {
                        "description": "Report updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}